feat(analysis): dump complete analysis reports
