+++ src/tilefoundry/dump/flags.py
@@ -4,7 +4,7 @@
 active.
 
 Initial set per `#tilefoundry:3beef367` 2026-04-27:
-``PASS_IR | CODEGEN_SOURCE | BUILD_LOG``, plus ``ALL`` as their union.
+``PASS_IR | CODEGEN_SOURCE | BUILD_LOG | ANALYSIS``, plus ``ALL`` as their union.
 Finer per-pass flags are out of scope for the MVP.
 """
 from __future__ import annotations
@@ -17,7 +17,8 @@ class DumpFlags(IntFlag):
     PASS_IR = 1 << 0
     CODEGEN_SOURCE = 1 << 1
     BUILD_LOG = 1 << 2
-    ALL = PASS_IR | CODEGEN_SOURCE | BUILD_LOG
+    ANALYSIS = 1 << 3
+    ALL = PASS_IR | CODEGEN_SOURCE | BUILD_LOG | ANALYSIS
 
 
 __all__ = ["DumpFlags"]
